requirements.txt
# Core dependencies for dataset processing and validation
pydantic>=2.0.0
pillow>=9.0.0
numpy>=1.21.0
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.20.0

# Data handling and processing
pandas>=1.5.0
pyyaml>=6.0
tqdm>=4.64.0

# Computer vision and image processing
opencv-python>=4.8.0

# Development and testing
pytest>=7.0.0
black>=22.0.0
flake8>=5.0.0

# Visualization (optional)
matplotlib>=3.5.0
seaborn>=0.11.0 